Balls to approx roots and fmpz for minpolys
Big change: AlgebraicNumber is not parametric.
We use the Flint fmpz integer type for the coefficients
of the minimal polynomial.
We use balls and balls arithmetic implemented in Arblib
to represent the root of the minimal polynomial.
Type convert code simplifies quite a lot.
Incorporate certificate for the ball containing a unique
root of the minimal polynomial.